+++ test/ssh/test_server.py
@@ -9,7 +9,7 @@ except ImportError:
 from cryptoparser.common.exception import NotEnoughData
 
 from cryptoparser.ssh.record import SshRecordInit
-from cryptoparser.ssh.subprotocol import SshDisconnectMessage, SshProtocolMessage
+from cryptoparser.ssh.subprotocol import SshDisconnectMessage, SshProtocolMessage, SshSoftwareVersionUnparsed
 from cryptoparser.ssh.version import SshProtocolVersion, SshVersion
 
 from cryptolyzer.common.transfer import L4ClientTCP
@@ -37,7 +37,7 @@ class TestSshServerDefaults(TestL7ServerBase):
         threaded_server.start()
 
         result = self.get_result('localhost', threaded_server.l7_server.l4_transfer.bind_port)
-        self.assertEqual(result.versions, [SshProtocolVersion(SshVersion.SSH2), ])
+        self.assertEqual(result.protocol_versions, [SshProtocolVersion(SshVersion.SSH2), ])
 
 
 class TestSshServerHandshake(TestL7ServerBase):
@@ -64,7 +64,10 @@ class TestSshServerHandshake(TestL7ServerBase):
         l4_client = L4ClientTCP('localhost', self.threaded_server.l7_server.l4_transfer.bind_port)
         l4_client.init_connection()
 
-        l4_client.send(SshProtocolMessage(SshProtocolVersion(SshVersion.SSH2, 0), 'software_version').compose())
+        l4_client.send(SshProtocolMessage(
+            SshProtocolVersion(SshVersion.SSH2, 0),
+            SshSoftwareVersionUnparsed('software_version')
+        ).compose())
         l4_client.send(SshRecordInit(SshDisconnectMessage(SshReasonCode.BY_APPLICATION, 'by application')).compose())
         SshRecordInit(SshDisconnectMessage(SshReasonCode.PROTOCOL_ERROR, 'protocol error', 'en')).compose()
         self._receive_as_many_as_possibe_and_close(l4_client)
@@ -72,10 +75,13 @@ class TestSshServerHandshake(TestL7ServerBase):
     def test_error_invalid_message(self):
         l4_client = L4ClientTCP('localhost', self.threaded_server.l7_server.l4_transfer.bind_port)
         l4_client.init_connection()
-        l4_client.send(SshProtocolMessage(SshProtocolVersion(SshVersion.SSH2, 0), 'software_version').compose())
+        l4_client.send(SshProtocolMessage(
+            SshProtocolVersion(SshVersion.SSH2, 0),
+            SshSoftwareVersionUnparsed('software_version')
+        ).compose())
         l4_client.send(b'\x00\x00\x00\x05\x00\x00\x00\x00\xff')
         self._receive_as_many_as_possibe_and_close(l4_client)
 
     def test_ssh_client(self):
         result = self.get_result('localhost', self.threaded_server.l7_server.l4_transfer.bind_port)
-        self.assertEqual(result.versions, [SshProtocolVersion(SshVersion.SSH2, )])
+        self.assertEqual(result.protocol_versions, [SshProtocolVersion(SshVersion.SSH2, )])