+++ test/ssh/test_versions.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from cryptoparser.ssh.version import SshProtocolVersion, SshVersion
-
-from cryptolyzer.ssh.client import L7ClientSsh
+from cryptolyzer.ssh.client import L7ClientSsh, SshProtocolMessageDefault
 from cryptolyzer.ssh.server import L7ServerSsh
 from cryptolyzer.ssh.versions import AnalyzerVersions
 
@@ -20,8 +18,10 @@ class TestL7ClientBase(TestSshCases.TestSshClientBase):
 
 class TestSshVersions(TestL7ClientBase):
     def test_versions(self):
+        protocol_message_default = SshProtocolMessageDefault()
         threaded_server = L7ServerSshTest(L7ServerSsh('localhost', 0, timeout=0.2))
         threaded_server.start()
 
         result = self.get_result('localhost', threaded_server.l7_server.l4_transfer.bind_port)
-        self.assertEqual(result.versions, [SshProtocolVersion(SshVersion.SSH2, )])
+        self.assertEqual(result.protocol_versions, [protocol_message_default.protocol_version, ])
+        self.assertEqual(result.software_version, protocol_message_default.software_version)